fix datatype

+++ src/qilib/utils/python_json_structure.py
@@ -27,7 +27,7 @@ from qilib.utils.type_aliases import PJSValues
 class PythonJsonStructure(dict):  # type: ignore
     __serializable_key_types = (str, int)
     __serializable_python_types = (bool, int, float, complex, str, bytes)
-    __serializable_numpy_types = (np.float32, np.float64, np.int32, np.int64, np.cfloat)
+    __serializable_numpy_types = (np.float32, np.float64, np.int32, np.int64, np.complex128)
     __serializable_value_types = (*__serializable_python_types, *__serializable_numpy_types)
 
     def __init__(self, *args: Dict[str, PJSValues], **kwargs: Any) -> None:
